pyproject.toml
[project]
name = "job_search_agent"
version = "0.1.0"
description = "job-search-agent using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.117.0,<1.0.0",
    "google-generativeai>=0.8.5",
    "python-docx>=1.2.0",
]

[project.scripts]
job_search_agent = "job_search_agent.main:run"
run_crew = "job_search_agent.main:run"
train = "job_search_agent.main:train"
replay = "job_search_agent.main:replay"
test = "job_search_agent.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
